+++ CHANGELOG.md
@@ -9,6 +9,11 @@ ClickHouse Connect has been included as an official Apache Superset database con
 However, if you need compatibility with older versions of Superset, you may need clickhouse-connect
 v0.5.25, which dynamically loads the EngineSpec from the clickhouse-connect project.
 
+## UNRELEASED
+### Improvements
+- Added type `@overload`s for the `raw_query` method of `Client` for correct inference of the return
+  type based on the `stream: bool` argument. Thanks for [Martijn Thé](https://github.com/martijnthe) for the PR!
+
 ## 0.7.5, 2024-03-28
 ### Bug Fixes
 - Fixed client side binding for Python format strings using `%d` (int) and `%f` (float) format patterns.  Closes